incorporate PC port

+++ src/pokemon_size_record.c
@@ -77,9 +77,9 @@ static u8 TranslateBigMonSizeTableIndex(u16 a)
 
 static u32 GetMonSize(u16 species, u16 b)
 {
-    u64 unk2;
-    u64 unk4;
-    u64 unk0;
+    u32 unk2;
+    u32 unk4;
+    u32 unk0;
     u32 height;
     u32 var;
